+++ db_edu_util/db_edu_util/__init__.py
@@ -2,7 +2,7 @@
 Utility library used by build tools.
 """
 
-VERSION = '1.2.0'
+VERSION = '1.3.0'
 
 from typing import Callable, Iterable, Any
 from textwrap import TextWrapper